+++ packages/ui/src/mosaic/components/heading/heading.tsx
@@ -5,9 +5,10 @@ import React from 'react';
 import type { MosaicComponentProps } from '../../props';
 import { mergeStyleProps, themeProps } from '../../props';
 import { useContextProps } from '../../utils/context';
+import { reset } from '../reset.styles';
 import type { TypographyColor, TypographySize } from '../typography.styles';
-import { colors, sizes } from '../typography.styles';
-import { styles } from './heading.styles';
+import { colors, sizes, styles as typographyStyles } from '../typography.styles';
+import { styles as headingStyles } from './heading.styles';
 
 export interface HeadingProps extends MosaicComponentProps<'h2'> {
   size?: TypographySize;
@@ -33,7 +34,7 @@ export const Heading = React.forwardRef<HTMLHeadingElement, HeadingProps>(functi
   const props = {
     ...mergeStyleProps(
       themeProps('heading', { size, color }),
-      stylex.props(styles.base, sizes[size], colors[color]),
+      stylex.props(reset.base, typographyStyles.base, headingStyles.base, sizes[size], colors[color]),
       className,
       style,
     ),